py_scripts/requirements.txt
python-dotenv
requests

